requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic-settings==2.2.1
email-validator>=2.1.1
sqlalchemy==2.0.29
psycopg[binary]==3.1.18

motor==3.4.0